+++ ng_trajectory/criterions/profile/main.py
@@ -10,9 +10,18 @@ import sys, numpy
 
 from . import profiler
 
-from ng_trajectory.interpolators.utils import pointDistance
+from ng_trajectory.interpolators.utils import pointDistance, trajectoryClosestIndex
 
 import ng_trajectory.plot as ngplot
+
+from multiprocessing import Queue
+
+
+# Global variables
+CENTERLINE = None
+REFERENCE_PROGRESS = None
+OVERTAKING_POINTS = None
+
 
 # Parameters
 from ng_trajectory.parameter import *
@@ -49,11 +58,14 @@ P.createAdd("plot_timelines_width", 0.6, float, "Linewidth of the timelines. 0 =
 
 def init(**kwargs) -> None:
     """Initialize criterion."""
-    global REFERENCE
+    global REFERENCE, CENTERLINE, OVERTAKING_POINTS
 
     profiler.parametersSet(**kwargs)
 
     P.updateAll(kwargs)
+
+    OVERTAKING_POINTS = Queue()
+
 
     if P.getValue("save_solution_csv") == "$":
         P.update("save_solution_csv", kwargs.get("logfile").name + ".csv")
@@ -79,7 +91,7 @@ def compute(points: numpy.ndarray, overlap: int = None, penalty: float = 100.0, 
     t -- time of reaching the last point of the trajectory, [s], float
          minimization criterion
     """
-    global REFERENCE
+    global REFERENCE, CENTERLINE, REFERENCE_PROGRESS, OVERTAKING_POINTS
 
     if overlap is None:
         overlap = P.getValue("overlap")
@@ -89,6 +101,7 @@ def compute(points: numpy.ndarray, overlap: int = None, penalty: float = 100.0, 
     )
 
     invalid_points = []
+    closest_indices = []
 
     if REFERENCE is not None:
         _d = P.getValue("reference_dist")
@@ -109,6 +122,8 @@ def compute(points: numpy.ndarray, overlap: int = None, penalty: float = 100.0, 
                     __t += _t[-1]
                 else:
                     break
+
+            closest_indices.append(_ci)
 
             # If the points are too close to each other, return penalty
             if pointDistance([rx, ry], points[_ci, :]) < _d:
@@ -186,6 +201,27 @@ def compute(points: numpy.ndarray, overlap: int = None, penalty: float = 100.0, 
 
     if len(invalid_points) > 0:
         return float(penalty)
+
+
+    # Locate points where overtaking occurs
+    # Centerline is used to obtain track progression.
+    if REFERENCE is not None and CENTERLINE is not None:
+        if REFERENCE_PROGRESS is None:
+            REFERENCE_PROGRESS = [
+                trajectoryClosestIndex(CENTERLINE, REFERENCE[_i, :2])
+                for _i in range(len(REFERENCE))
+            ]
+
+        overtaken = False
+        for _i, (rx, ry, _) in enumerate(REFERENCE):
+            rd = REFERENCE_PROGRESS[_i]                    # nejblizsi i ve stejnem case
+            pd = trajectoryClosestIndex(CENTERLINE, points[closest_indices[_i], :2])
+
+            if rd > 50 and pd > rd and not overtaken:
+                overtaken = True
+                OVERTAKING_POINTS.put(points[closest_indices[_i], :2])
+            #elif pd < rd and overtaken:
+            #    overtaken = False
 
 
     return float(_t[-1])